osfunc: StartProcess: generate report and exit properly on exception
Previously only the exception would be logged with LLError. Now the
Exception message is logged as LLDebug (which maybe should be changed
somehow?) and Report is set to a message informing about the error and
Result is false.

This way functions calling StartProcess a) won't erroneously believe
that StartProcess was successful and b) will have to do error handling
completely themselves (which is already in place in the relevant
locations).