osparser: executeWith: don't check syntax twice
Now we only check for compatibility inside executeWith and let
produceExecLine do the syntax checking on its own. This double
checking was not necessary at all and led to options being missing
in one part and thus breaking in cases it shouldn't.

Now the behavior is similar to the original ExecWith behavior except
it generates an error if a Windows-only option is used on non-Windows.
This is the case for /64bit /SysNative /32bit and all /Run-Parameters.