osparser: shellCall: improve performance
While looking for "cmd64.exe.mui" files we now only search
direct subdirectories of "C:\windows" which is where those files are
located instead of all subdirectories of all subdirectories recursively.
This improves the performance of shellCall greatly. The poor performance
was an issue previously, as it caused the processCall
/WaitForWindowAppearing test in opsi-script-test to fail.

To achieve this a new function was added to osfunc.pas FindInSubDirs.