osparser: improve handling of windows-only params
* executeWith: only log error, continue like normal
* execDosBatch: log message as error

These changes are to make handling simpler and more in
line to previous behavior.
In the case of ExecWith previously Windows-Only parameters
where just ignored. This behavior has been restored.
A recent change to DosBatch also causes it to not fail on
Syntax Errors, so we just log any occurring errors.