osfunc: SetFilePermissionForRunAs: no traPcpatch
Disable this feature since it is unused anyways in opsi-script, so there
is no intended way to trigger this code path. As it can't be expected
that a pcpatch account is present, the implementation makes no real
sense, so it would need to be reworked if it was ever used.

Therefore it currently makes no sense to keep it around. Instead we just
generate an error with a appropriate log message, whenever
SetFilePermissionForRunAs is called with traPcpatch.