osfunc: StartProcess: don't catch out if threaded
This change ensures that processes don't have a proper
StdOut or StdErr if started threaded since poUsePipes
or STARTF_USESTDHANDLES are not set respectively.
Also, if CreateProcess is used directly, we disable
handle inheritance additionally.

Not doing this could break some processes, since they
would write to a pipe that would never be read. This
especially affected helperchild.exe of opsi-script-test
which broke a test, since the program behaves completely
differently to its options if started with /LetThemGo
prior to this change.